+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ThrownExperienceBottleMixin.java
@@ -1,18 +1,22 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import com.llamalad7.mixinextras.expression.Definition;
+import com.llamalad7.mixinextras.expression.Expression;
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.iguanatweaksreborn.module.experience.DroppedExperience;
 import net.minecraft.world.entity.projectile.ThrownExperienceBottle;
-import net.minecraft.world.phys.HitResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ThrownExperienceBottle.class)
 public class ThrownExperienceBottleMixin {
 
-	@Inject(at = @At("HEAD"), method = "onHit")
-	public void onHit(HitResult hitResult, CallbackInfo callbackInfo) {
-		DroppedExperience.onXpBottleHit((ThrownExperienceBottle) (Object) this);
+	@Definition(id = "level", method = "Lnet/minecraft/world/entity/projectile/ThrownExperienceBottle;level()Lnet/minecraft/world/level/Level;")
+	@Definition(id = "random", field = "Lnet/minecraft/world/level/Level;random:Lnet/minecraft/util/RandomSource;")
+	@Definition(id = "nextInt", method = "Lnet/minecraft/util/RandomSource;nextInt(I)I")
+	@Expression("? + this.level().random.nextInt(?) + this.level().random.nextInt(?)")
+	@ModifyExpressionValue(method = "onHit", at = @At("MIXINEXTRAS:EXPRESSION"))
+	public int onHit(int original) {
+		return DroppedExperience.getXpBottleDroppedExperience((ThrownExperienceBottle) (Object) this);
 	}
 }