+++ changelog.md
@@ -14,6 +14,7 @@
   * Increased xp to level up (35 -> 50)
   * Reduced materials enchantability by ~25%
   * Unbreaking max level has been reduced from 5 to 3 and increased durability from 50% to 75%
+  * Bottle o' Enchanting now always award 50 xp (1 level)
 * Removed shields from zombies
 * Cocoa beans now drop 2 beans instead of 3
 * Increased crossbow % damage (-50% -> -40%)